[BUGFIX] Change of time, longitude and latitude types from int to float

+++ components/common/data/src/AccelerometerSensorData.hpp
@@ -17,7 +17,7 @@ public:
     float acceleration_z = 0.0f;
     
     // Metadata
-    uint32_t timestamp = 0;
+    float timestamp = 0;
 
     json toJSON() const override {
         json j;
